pyproject.toml
[tool.poetry]
name = "FastAPI-Traefik"
version = "0.1.0"
description = ""
authors = ["GGGoingdown <eddiepai0401@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
dependency-injector = "^4.39.1"
sentry-sdk = "^1.9.0"
loguru = "^0.6.0"
ecs-logging = "^2.0.0"
shortuuid = "^1.0.9"
aerich = "^0.6.3"
tortoise-orm = "0.19.0"
aioredis = "^2.0.1"
asyncpg = "^0.26.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^1.9.1"}
tenacity = "^8.0.1"
websockets = "^10.3"
python-socketio = "^5.7.2"

[tool.poetry.dev-dependencies]
pytest = "7.1.1"
pytest-asyncio = "0.18.3"
httpx = "0.22"
requests = "^2.28.1"


[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
